+++ src/test/java/org/prgms/TestConfig.java
@@ -1,0 +1,29 @@
+package org.prgms;
+
+import com.zaxxer.hikari.HikariDataSource;
+import org.springframework.boot.jdbc.DataSourceBuilder;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.jdbc.core.JdbcTemplate;
+
+import javax.sql.DataSource;
+
+@TestConfiguration
+@ComponentScan(basePackages = {"org.prgms.voucher"})
+public class TestConfig {
+    @Bean
+    public DataSource dataSource() {
+        return DataSourceBuilder.create()
+                .url("jdbc:mysql://localhost:2215/test-order_mgmt")
+                .username("test")
+                .password("test1234!")
+                .type(HikariDataSource.class)
+                .build();
+    }
+
+    @Bean
+    public JdbcTemplate jdbcTemplate(DataSource dataSource) {
+        return new JdbcTemplate(dataSource);
+    }
+}
